Refactor financial data types to use double instead of Decimal
- Updated the `Kline` and `Ohlcv` classes to use `double` for price and volume fields, enhancing performance and simplifying calculations.
- Modified the `DecList` type definition to reflect the change from `Decimal` to `double`.
- Removed the `Decimal` dependency from the project by deleting the `ohlcv_series_test.dart` file, which was reliant on it.
- Adjusted related methods and serialization logic to accommodate the new data types.

+++ lib/src/type/ohlcv/ohlcv.dart
@@ -9,11 +9,11 @@ part 'ohlcv.g.dart';
 @freezed
 abstract class Ohlcv with _$Ohlcv {
   factory Ohlcv({
-    required Decimal open,
-    required Decimal high,
-    required Decimal low,
-    required Decimal close,
-    required Decimal volume,
+    required double open,
+    required double high,
+    required double low,
+    required double close,
+    required double volume,
     required int openTimestamp,
     required int closeTimestamp,
   }) = _Ohlcv;
@@ -22,7 +22,7 @@ abstract class Ohlcv with _$Ohlcv {
 
   const Ohlcv._();
 
-  Decimal price(OhlcvType type) {
+  double price(OhlcvType type) {
     switch (type) {
       case OhlcvType.open:
         return open;
